requirements.txt
aioch==0.0.2
aiogram==2.23.1
aiohttp==3.8.3
aiokafka==0.7.2
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
asyncpg==0.27.0
atlastk==0.13.2
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
clickhouse-driver==0.2.4
dacite==1.6.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
kafka-python==2.0.2
magic-filter==1.0.9
multidict==6.0.2
pypika-tortoise==0.1.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
tortoise-orm==0.19.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
yarl==1.8.1

Pillow==9.3.0
